parser: section getKey handle sections with same name
Previously, the section getKey would return keys for the first section
found. If there is more than one section with the same name, the first
section was always used when searching for keys in those sections. This
is because it loops through all the file lines looking for a section
with the same name, and the first one declared was always foudn first.

Use the section position, rather than name, to determine the section we
are looking for keys for.